+++ core/datasources-service/tests/create.test.js
@@ -3,7 +3,8 @@ const fse = require('fs-extra');
 const HttpStatus = require('http-status-codes');
 const { uid: uuid } = require('@hkube/uid');
 const validationMessages = require('../lib/consts/validationMessages.js');
-const { createDataSource, mockRemove } = require('./utils');
+const { createDataSource } = require('./api');
+const { mockRemove, mockDeleteClone } = require('./utils');
 
 let DATASOURCE_GIT_REPOS_DIR;
 
@@ -98,15 +99,12 @@ describe('/datasource POST', () => {
     describe('create', () => {
         it("should create a new dataSource and return it's newly created id and files list", async () => {
             const name = uuid();
-            const removeDirectoryMock = mockRemove();
+            const deleteClone = mockDeleteClone();
             const {
                 response: { statusCode },
                 body: dataSource,
             } = await createDataSource({ body: { name } });
-
-            expect(removeDirectoryMock.getCall(0).firstArg).to.match(
-                new RegExp(name)
-            );
+            expect(deleteClone.getCalls()).to.have.lengthOf(1);
             expect(statusCode).to.eql(HttpStatus.CREATED);
             expect(dataSource).to.have.property('id');
             expect(dataSource).to.have.property('name');
@@ -119,6 +117,11 @@ describe('/datasource POST', () => {
                 'utf8'
             );
             expect(JSON.parse(hkubeFile)).to.eql({ repositoryName: name });
+            const config = fse.readFileSync(
+                `${DATASOURCE_GIT_REPOS_DIR}/${name}/.dvc/config.local`,
+                'utf8'
+            );
+            expect(config).to.match(new RegExp(name));
         });
         it('should create an empty dataSource', async () => {
             const name = uuid();
@@ -148,16 +151,6 @@ describe('/datasource POST', () => {
             expect(secondResponse.body.error.message).to.contain(
                 'already exists'
             );
-        });
-        it('should configure storage for the datasource', async () => {
-            const name = uuid();
-            mockRemove();
-            await createDataSource({ body: { name } });
-            const config = fse.readFileSync(
-                `${DATASOURCE_GIT_REPOS_DIR}/${name}/.dvc/config.local`,
-                'utf8'
-            );
-            expect(config).to.match(new RegExp(name));
         });
         it('should upload a file with meta data', async () => {
             const name = uuid();
