+++ core/datasources-service/tests/dataSource.test.js
@@ -5,15 +5,14 @@ const { uid: uuid } = require('@hkube/uid');
 const sinon = require('sinon');
 const { request } = require('./request');
 const dbConnection = require('../lib/db');
+const { nonExistingId, mockDeleteClone } = require('./utils');
 const {
     createDataSource,
     fetchDataSource,
     deleteDataSource,
     updateVersion,
-    nonExistingId,
     fetchDataSourceVersions,
-    mockRemove,
-} = require('./utils');
+} = require('./api');
 
 let DATASOURCE_GIT_REPOS_DIR, restUrl, restPath;
 
@@ -215,11 +214,11 @@ describe('Datasource', () => {
     describe('/datasource/:name/versions GET', () => {
         it('should fetch the versions listing of a datasource', async () => {
             const name = uuid();
-            const mockedRemove = mockRemove();
+            const deleteClone = mockDeleteClone();
             const { body: dataSource } = await createDataSource({
                 body: { name },
             });
-            expect(mockedRemove.callCount).to.be.gt(1);
+            expect(deleteClone.callCount).to.eq(1);
             const [existingFile] = dataSource.files;
             expect(
                 await fse.pathExists(
@@ -230,7 +229,6 @@ describe('Datasource', () => {
                 `${DATASOURCE_GIT_REPOS_DIR}/${name}/data/${existingFile.name}`
             );
             expect(fileStats.size).to.eq(107);
-            mockedRemove.resetHistory();
             const uploadResponse = await updateVersion({
                 dataSourceName: name,
                 fileNames: ['updatedVersions/README-1.md'],
@@ -239,7 +237,7 @@ describe('Datasource', () => {
             const { body: versionsList } = await fetchDataSourceVersions({
                 name,
             });
-            expect(mockedRemove.callCount).to.be.gt(1);
+            expect(deleteClone.callCount).to.eq(2);
             expect(versionsList).to.have.lengthOf(2);
             // validates the order of the array
             const [, latestVersion] = versionsList;
